Adjustments made after testing.
