+++ src/Sanctuary.Game/Entities/IEntity.cs
@@ -33,7 +33,7 @@ public interface IEntity : IEquatable<IEntity>, IDisposable
     #region Update
 
     void UpdatePosition(Vector4 position, Quaternion rotation, bool updateZoneArea = true);
-    void TeleportToZone(IZone zone, Vector4 position, Quaternion rotation);
+    bool TeleportToZone(IZone zone, Vector4 position, Quaternion rotation);
 
     void UpdateEveryTick();
     void UpdateEverySecond();
