+++ src/Sanctuary.Game/Entities/Mount.cs
@@ -21,7 +21,7 @@ public class Mount : Npc
         Definition = definition;
     }
 
-    public override void TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
+    public override bool TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
     {
         // Alert/Remove visible entities
         foreach (var visiblePlayer in VisiblePlayers)
@@ -46,6 +46,8 @@ public class Mount : Npc
         ZoneTile = ZoneTile.Empty;
 
         UpdatePosition(position, rotation);
+
+        return true;
     }
 
     public override PlayerUpdatePacketAddNpc GetAddNpcPacket()
