+++ src/Sanctuary.Game/Entities/Npc.cs
@@ -172,8 +172,9 @@ public class Npc : IScriptableNpc, IEntity
         }
     }
 
-    public virtual void TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
+    public virtual bool TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
     {
+        return true;
     }
 
     protected void UpdateZoneTile()
