+++ src/Sanctuary.Game/Entities/Player.cs
@@ -9,6 +9,7 @@ using System.Threading.Tasks;
 
 using Sanctuary.Core.Collections;
 using Sanctuary.Core.IO;
+using Sanctuary.Game;
 using Sanctuary.Game.ChatCommands;
 using Sanctuary.Game.Helpers;
 using Sanctuary.Game.Interactions;
@@ -25,6 +26,7 @@ public sealed class Player : ClientPcData, IEntity
 {
     private readonly UdpConnection _connection;
     private readonly IResourceManager _resourceManager;
+    private readonly IZoneManager _zoneManager;
 
     public bool Visible { get; set; }
 
@@ -67,12 +69,13 @@ public sealed class Player : ClientPcData, IEntity
     public Vector4 StartingZonePosition { get; set; }
     public Quaternion StartingZoneRotation { get; set; }
 
-    public Player(BaseZone zone, UdpConnection connection, IResourceManager resourceManager)
+    public Player(BaseZone zone, UdpConnection connection, IResourceManager resourceManager, IZoneManager zoneManager)
     {
         Zone = zone;
 
         _connection = connection;
         _resourceManager = resourceManager;
+        _zoneManager = zoneManager;
     }
 
     #region Connection
@@ -209,19 +212,24 @@ public sealed class Player : ClientPcData, IEntity
         ZoneTile = newZoneTile;
     }
 
-    public void TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
+    public bool TeleportToZone(IZone zone, Vector4 position, Quaternion rotation)
     {
         if (Zone == zone)
-            return;
+            return true;
+
+        if (zone.IsDisposed)
+        {
+            if (!_zoneManager.TryGetOrCreateZoneInstance(zone.DefinitionId, zone.OwnerId, out var freshZone))
+                return false;
+
+            zone = freshZone;
+        }
 
         if (Zone is WorldZone)
         {
             StartingZonePosition = Position;
             StartingZoneRotation = Rotation;
         }
-
-        if (Mount is not null)
-            Mount.TeleportToZone(zone, position, rotation);
 
         // Alert/Remove visible entities
         foreach (var visiblePlayer in VisiblePlayers)
@@ -232,11 +240,20 @@ public sealed class Player : ClientPcData, IEntity
 
         ZoneTile.Entities.Remove(Guid, out _);
 
-        Zone.TryRemovePlayer(Guid);
+        var oldZone = Zone;
+
+        oldZone.TryRemovePlayer(Guid);
 
         // Add to new zone/zonetile
 
-        zone.TryAddPlayer(this);
+        if (!zone.TryAddPlayer(this))
+        {
+            oldZone.TryAddPlayer(this);
+            return false;
+        }
+
+        if (Mount is not null)
+            Mount.TeleportToZone(zone, position, rotation);
 
         // Teleport to new zone
 
@@ -253,13 +270,15 @@ public sealed class Player : ClientPcData, IEntity
             Name = Zone.Name,
             Position = position,
             Rotation = rotation,
-            Sky = "sky_deep_mines.xml",
+            Sky = null,
             Id = Zone.Id,
             GeometryId = 214,
             OverrideUpdateRadius = true
         };
 
         SendTunneled(packetClientBeginZoning);
+
+        return true;
     }
 
     private void UpdateZoneArea()
