+++ src/Sanctuary.Game/IZoneManager.cs
@@ -1,4 +1,5 @@
-﻿using System.Diagnostics.CodeAnalysis;
+﻿using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 
 using Sanctuary.Game.Entities;
 using Sanctuary.Game.Zones;
@@ -9,8 +10,15 @@ public interface IZoneManager
 {
     WorldZone StartingZone { get; }
 
+    IEnumerable<IZone> Zones { get; }
+
     bool Load();
 
     bool TryGetPlayer(ulong guid, [MaybeNullWhen(false)] out Player player);
     bool TryGetPlayer(string name, [MaybeNullWhen(false)] out Player player);
+
+    bool TryGetOrCreateZoneInstance(int zoneDefinitionId, ulong? ownerId, [MaybeNullWhen(false)] out IZone zone);
+    bool TryGrantHouse(int zoneDefinitionId, ulong ownerId);
+
+    void RemoveZoneInstance(IZone zone);
 }
